Bugfixes & enhancements, new planting date raster
New planting date raster has been generated using CRU and EcoCrop. This
is much closer to the original version that was used in the analysis
found in Global Change Biology.

Correction in the function name in Function/Get_CRU_20_data.R

Removal of some code to parameterise the EcoCrop model such that the
planting seasons are now more closely aligned with the original data
that was used. There have been changes in the R package for EcoCrop
since the original work was done six years ago. This recreates the
planting dates as closely as possible to the original data.