Updates and completion
Finalise this repository.

All data used herein are freely available and automatically downloaded
and imported into R using these scripts.

Maps of average country change and change in Nepal are now presented
along with a graph of the top 10 countries experiencing the most change
and the change in the top 10 potato producing countries.

FAO production data now defaults to the most recent year of data
provided.

Scripts mostly follow Wickham’s style guide.

Directories and scripts are numbered sequentially in order in which
they were originally and should be run to generate necessary data.

Used the readr package in this iteration to greatly increase the speed
at which data is imported (mainly FAO data).

Other enhancements, and cleaning abound.